+++ create_db/wiktionary_de_parser/__init__.py
@@ -25,8 +25,9 @@ class Parser:
         methods_path = os.path.join(os.path.dirname(__file__), 'methods')
         method_files = [f for f in os.listdir(methods_path) if not f.startswith('__') and f.endswith('.py')]
         for idx, f in enumerate(method_files):
-            full_path = os.path.join(methods_path, f)
-            module = SourceFileLoader('method-' + str(idx), full_path).load_module()
+            fullpath = os.path.join(methods_path, f)
+            fullname = 'method-' + str(idx)
+            module = SourceFileLoader(fullname=fullname, path=fullpath).load_module()
 
             if not hasattr(module, 'init'):
                 raise Exception(f'No init() method found in file "{f}"')
@@ -82,9 +83,9 @@ class Parser:
 
     def __iter__(self):
         """
-        Iterate all '<tag>...</tag>' Element Trees yielded from self._parse()
+        Iterate all pages yielded from self.parse_page() and all word sections yielded from self.parse_sections()
 
-        :return: Dict var 'entity' {tag1, value, tag2, value, ... ,tagn, value}}
+        :return: Dict with final result
         """
         for title, wikitext in self.parse_page():
             # check for ignored titles