+++ create_db/wiktionary_de_parser/methods/pos.py
@@ -106,7 +106,9 @@ pos_map = {
         'Erweiterter Infinitiv',
     ],
     'Wortverbindung': [],
-    'Flektierte Form': [],
+
+    # needs additional parsing; can be: Substantiv, Adjektiv, Artikel, Pronomen
+    'Deklinierte Form': []
 }
 
 if debug is True:
@@ -115,42 +117,34 @@ if debug is True:
     all_pos_names = [x.lower() for x in all_pos_names]
 
 
-def find_pos(title, pos_names, text):
+def find_pos(title, pos_names, text, current_record):
     result = {}
-
-    # check for "Deklinierte Form" first
-    # can be both: Substantiv/Adjektiv
-    if 'Deklinierte Form' in pos_names:
-        # get first uppercase letter
-        upper_chars = []
-        for char in title:
-            if char.isupper():
-                upper_chars.append(char)
-                break
-        if upper_chars:
-            result['Substantiv'] = ['Deklinierte Form']
-        else:
-            result['Adjektiv'] = ['Deklinierte Form']
-        # remove from names
-        del pos_names[pos_names.index('Deklinierte Form')]
 
     # fix POS when there is a certain POS template, but POS is not in pos_names
     # example "Substantiv": https://de.wiktionary.org/wiki/wei%C3%9Fes_Gold
-    if 'Substantiv' not in result:
-        if '{{Deutsch adjektivisch Übersicht' in text \
-                or '{{Deutsch Substantiv Übersicht - sch' in text \
-                or '{{Deutsch Substantiv Übersicht' in text:
+
+    # Substantiv
+    if '{{Deutsch adjektivisch Übersicht' in text \
+            or '{{Deutsch Substantiv Übersicht - sch' in text \
+            or '{{Deutsch Substantiv Übersicht' in text \
+            or '{{Deutsch Toponym Übersicht' in text:
+        if 'Substantiv' not in result:
             result['Substantiv'] = []
-    if 'Adjektiv' not in result:
-        if '{{Deutsch Adjektiv Übersicht' in text:
-            result['Adjektiv'] = []
-    if '{{Deutsch Adverb Übersicht' in text:
+    if '{{Deutsch adjektivisch Übersicht' in text and 'adjektivische Deklination' not in result['Substantiv']:
+        result['Substantiv'].append('adjektivische Deklination')
+    if '{{Deutsch Toponym Übersicht' in text and 'Toponym' not in result['Substantiv']:
+        result['Substantiv'].append('Toponym')
+    # Adjektiv
+    if '{{Deutsch Adjektiv Übersicht' in text and 'Adjektiv' not in result:
+        result['Adjektiv'] = []
+    # Adverb
+    if '{{Deutsch Adverb Übersicht' in text and 'Adverb' not in result:
         result['Adverb'] = []
-    if '{{Deutsch Pronomen Übersicht' in text:
+    # Pronomen
+    if '{{Deutsch Pronomen Übersicht' in text and 'Pronomen' not in result:
         result['Pronomen'] = []
-    if '{{Deutsch Toponym Übersicht' in text:
-        result['Toponym'] = []
-    if '{{Deutsch Verb Übersicht' in text:
+    # Verb
+    if '{{Deutsch Verb Übersicht' in text and 'Verb' not in result:
         result['Verb'] = []
 
     # map other pos names
@@ -163,10 +157,12 @@ def find_pos(title, pos_names, text):
                 result[key] = []
             values_low = [x.lower() for x in values]
             if name_low in values_low:
-                name_idx = values_low.index(name_low)
                 if key not in result:
                     result[key] = []
-                result[key].append(values[name_idx])
+                name_idx = values_low.index(name_low)
+                value = values[name_idx]
+                if value not in result[key]:
+                    result[key].append(value)
 
     if debug is True:
         not_found_names = [x for x in pos_names if x.lower() not in all_pos_names and x not in not_in_map]
@@ -190,10 +186,11 @@ def init(title, text, current_record):
     if not pos_names:
         return False
 
-    if not pos_names:
-        return False
+    # strip
+    pos_names = [name.strip() for name in pos_names]
 
-    pos_normalized = find_pos(title, pos_names, text)
+    # find in map
+    pos_normalized = find_pos(title, pos_names, text, current_record)
     if not pos_normalized.keys():
         return False